fixing backup user flow
backups flow now will first try and make a backup of the file that's about to be overwritten before writing changes to the file. The user can open the backups folder with menu to load a backup if there was an issue with the writes. TODO: add a rotation for backups to not fill backups folder.